pyproject.toml
[build-system]
requires = ["uv_build>=0.9.9,<0.10.0"]
build-backend = "uv_build"

[project]
name = "cryo-ttt"
version = "0.1.0"
description = "F2FD-TTT: Test-time training with self-supervised Fourier-to-Fourier denoising for membrane segmentation in cryo-ET tomograms."
readme = "README.md"
requires-python = "~=3.10.0"
license = "MIT"
license-files = ["LICEN[CS]E*"]
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Development Status :: 4 - Beta",
]
dependencies = [
    "h5py>=3.14.0",
    "hydra-core~=1.3.2",
    "matplotlib~=3.10.3",
    "monai~=1.4.0",
    "mrcfile~=1.5.4",
    "nibabel~=5.3.2",
    "pandas~=2.3.0",
    "pillow~=11.0.0",
    "pre-commit>=4.4.0",
    "pytorch-lightning~=2.3.0",
    "scipy~=1.15.3",
    "seaborn~=0.13.2",
    "torch==2.0.1",
    "torchaudio==2.0.2",
    "torchmetrics~=1.6.3",
    "torchvision==0.15.2",
    "wandb~=0.19.11",
]
authors = [
    {name="Diyor Khayrutdinov", email="dijor0310@gmail.com"}
]

[project.optional-dependencies]
dev = [
    "ipykernel>=6.29.5",
]

[project.urls]
"Repository" = "https://github.com/dijor0310/cryo-ttt"
"Homepage" = "https://github.com/dijor0310/cryo-ttt"
"Bug Tracker" = "https://github.com/dijor0310/cryo-ttt/issues"

[tool.uv.build-backend]
module-name = "ttt"
namespace = true

[dependency-groups]
dev = [
    "ruff>=0.14.5",
]
